requirements.txt
Django==5.2.5
psycopg2-binary==2.9.10
python-dotenv==1.1.1
opencv-python==4.12.0.88
deepface==0.0.95
numpy==2.2.6
Pillow==11.3.0
requests==2.32.4
twilio==9.8.3
pypdf==5.3.1
google-genai==0.8.0
groq==0.18.0

